Added "The Captain Is Dead".
"The Captain Is Dead" is a Stargate shuttle simulator for Roblox.

The Captain is dead. You are out of fuel. And you are lost in space. You must find out how the controls work in order to scavage the nearby crash-site for fuel and hyper-jump to the "Destiny" mother-ship. 

Do not jettison the door. Do not press the self-destruction button. And do not get flushed into space.